Attempt to organize the core engine boundary with mixed success
