+++ packages/voictents-and-estinants-engine/src/core/engine-shell/appreffinge/framation.ts
@@ -1,5 +1,5 @@
-import { ZornTuple } from '../utilities/semantic-types/zorn';
-import { Hubblepup } from './hubblepup';
+import { ZornTuple } from '../../../utilities/semantic-types/zorn';
+import { Hubblepup } from '../quirm/hubblepup';
 
 /**
  * A function that takes the leftmost Hubblepup of a Cology and outputs a Zorn tuple for one of the other members of the Cology
