+++ packages/voictents-and-estinants-engine/src/core/engine-shell/quirm/hubblepup.ts
@@ -1,4 +1,4 @@
-import { StralineObject } from '../utilities/semantic-types/straline';
+import { StralineObject } from '../../../utilities/semantic-types/straline';
 
 /**
  * A thing that a Concrete Programmer wants to operate on
