+++ packages/voictents-and-estinants-engine/src/core/internal/procody/procody.ts
@@ -1,4 +1,4 @@
-import { Gepp } from './gepp';
+import { Gepp } from '../../engine-shell/voictent/gepp';
 import { Ajorken } from './ajorken';
 
 /**
