--- packages/voictents-and-estinants-engine/src/core/tomoode.ts
@@ -1,5 +1,0 @@
-/**
- * A placeholder for things that have not been implemented yet
- */
-export const TOMOODE = Symbol('NOT IMPLEMENTED');
-export type Tomoode = typeof TOMOODE;
