+++ packages/voictents-and-estinants-engine/src/custom/programs/scaffold-voictent-file/assembleScaffoldedFile.ts
@@ -1,0 +1,23 @@
+import fs from 'fs';
+import { buildBasicQuirmDebugger } from '../../debugger/quirmDebugger';
+import { digikikify } from '../../../type-script-adapter/digikikify';
+import { SCAFFOLD_CONFIGURATION_GEPP } from './scaffoldConfiguration';
+import { scaffoldFile } from './scaffoldFile';
+
+const [filePath] = process.argv.slice(2);
+
+if (filePath === undefined) {
+  throw Error('filePath is required');
+}
+
+if (!fs.existsSync(filePath)) {
+  throw Error(`"${filePath}" does not exist`);
+}
+
+digikikify({
+  initialVoictentsByGepp: {
+    [SCAFFOLD_CONFIGURATION_GEPP]: [{ filePath }],
+  },
+  estinantTuple: [scaffoldFile],
+  quirmDebugger: buildBasicQuirmDebugger('assembleScaffoldedFile'),
+});
