+++ packages/voictents-and-estinants-engine/src/type-script-adapter/gepp.ts
@@ -1,3 +1,3 @@
-import { Gepp as CoreGepp } from '../core/gepp';
+import { Gepp as CoreGepp } from '../core/engine-shell/voictent/gepp';
 
 export type Gepp<TGepp extends CoreGepp = CoreGepp> = TGepp;
