+++ packages/voictents-and-estinants-engine/src/type-script-adapter/quirm.ts
@@ -1,4 +1,4 @@
-import { Quirm as CoreQuirm } from '../core/quirm';
+import { Quirm as CoreQuirm } from '../core/engine-shell/quirm/quirm';
 
 export type Quirm<TQuirm extends CoreQuirm = CoreQuirm> = TQuirm;
