--- packages/voictents-and-estinants-engine/src/utilities/passthrough.ts
@@ -1,1 +1,0 @@
-export const passthrough = <T>(input: T): T => input;
