--- packages/voictents-and-estinants-engine/src/utilities/typed-datum/type-script/primitive.ts
@@ -1,11 +1,0 @@
-export type TypeScriptPrimitiveOptionTuple = readonly [
-  bigint,
-  boolean,
-  null,
-  number,
-  string,
-  symbol,
-  undefined,
-];
-
-export type TypeScriptPrimitive = TypeScriptPrimitiveOptionTuple[number];
